pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Uyo-obong Akpan <uyoobonga@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytz = "^2022.7"
requests = "^2.28.1"
Celery= "5.2.7"
Django="4.1.4"
django-environ="0.9.0"
django-filter="22.1"
djangorestframework="3.14.0"
drf-yasg="1.21.4"
gunicorn="20.1.0"
psycopg2-binary="2.9.5"
redis="4.4.0"
sentry-sdk="1.12.1"
Pygments="2.13.0"
Werkzeug="2.2.2"
bandit="1.7.4"
black="22.12.0"
coverage="7.0.0"
django-debug-toolbar="3.8.1"
django-extensions="3.2.1"
docker-compose="1.29.2"
fabric="2.7.1"
invoke="1.7.3"
ipdb="0.13.11"
ipython="8.7.0"
isort="5.11.4"
model-bakery="1.9.0"
pip-tools="6.12.1"
pre-commit="2.20.0"
pylama="8.4.1"
pylint-celery="0.3"
pylint-django="2.5.3"
pylint="2.15.9"
pytest-cov="4.0.0"
pytest-django="4.5.2"
pytest-mock="3.10.0"
pytest="7.2.0"
radon="5.1.0"
safety="2.3.5"
djangorestframework-simplejwt = "^5.2.2"
PyJWT = "^2.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
